transform/dim_institution/requirements.txt
pandas==1.5.2
python-decouple==3.6
fastparquet==2023.2.0
pdbpp==0.10.3
pyarrow==10.0.1
